connectors-gateway: Rename moonbeam to ethereum_xcm
